requirements.txt
google-genai==1.20.0
streamlit==1.45.1